Fix: Make smoothing window value consistent
- mavenparameters->smoothingAlgorithm was used to save the value of smoothingWindow
- Initial value of smoothingWindow in settings was getting set to 5 instead of 10
- smoothingWindow was int and float in different parts of the code. int is consistently being used now

+++ src/gui/mzroll/eicwidget.cpp
@@ -1568,8 +1568,8 @@ void EicWidget::groupPeaks() {
 	//qDebug() << "EicWidget::groupPeaks() " << endl;
 	//delete previous set of pointers to groups
 	QSettings *settings = getMainWindow()->getSettings();
-	float eic_smoothingWindow =
-			settings->value("eic_smoothingWindow").toDouble();
+	int eic_smoothingWindow =
+			settings->value("eic_smoothingWindow").toInt();
 	float grouping_maxRtWindow =
 			settings->value("grouping_maxRtWindow").toDouble();
